requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Data Processing & Analytics
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
prophet==1.1.4
networkx==3.2.1

# Visualization & Dashboards
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-table==5.0.0
streamlit==1.28.2

# API & HTTP
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Database & Time Series
boto3==1.34.0
aioboto3==12.3.0

# Geospatial & Mapping
geopy==2.4.1
folium==0.15.1

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development & Utilities
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# AWS Services
boto3==1.34.0
aioboto3==12.3.0

# Additional ML Libraries
xgboost==2.0.2
lightgbm==4.1.0
optuna==3.5.0

# Time Series & Forecasting
statsmodels==0.14.0
arch==6.2.0

# Optimization
pulp==2.7.0
ortools==9.8.3296 